+++ tests/testutils.sh
@@ -30,6 +30,37 @@ sed_literal() {
     printf '%s\n' "$1" | sed -e 's/[]\/$*.^[]/\\&/g'
 }
 
+# used below to normalise machine names
+# shellcheck disable=SC2016
+_awk_clean_name='
+#!/bin/awk -f
+# consistently replace numbers in arg:NNN conArg:NNN and $resolvedNNN
+BEGIN { count = 1 }
+{
+    out = ""
+    while (match($0, /(arg:|conArg:|[$]resolved)[0-9]*/)) {
+        rs = RSTART
+        rl = RLENGTH
+        m = substr($0, rs, rl - 1)
+        pfx = "$resolved"
+        if (match(m,/arg:/)) { pfx = "arg:" }
+        if (match(m,/conArg:/)) { pfx = "conArg:" }
+        if (!(m in mapping)) {
+            mapping[m] = count
+            count++
+        }
+        out = out substr($0, 1, rs - 1) pfx mapping[m]
+        $0 = substr($0, rs + rl)
+    }
+    print out $0
+}
+'
+
+# normalise machine names
+clean_names() {
+    awk "$_awk_clean_name"
+}
+
 # Folder containing the currently running test
 if [ "$OS" = "windows" ]; then
     test_dir="$(cygpath -m "$(pwd)")"