+++ CHANGELOG.md
@@ -314,6 +314,7 @@
 * Updates the docs for `envvars` to categorise when environment variables are
   used (runtime, build-time, or both).
 * Fixed build failure occuring when `make -j` is in effect.
+* Add `clean_names` function to `testutils.sh` to normalise machine names
 
 ## v0.6.0